% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-run.R
\name{compare_runs}
\alias{compare_runs}
\title{compare runs}
\usage{
compare_runs(
  id1,
  id2,
  output_dir = file.path(tempdir(), sprintf("\%s-\%s", id1, id2)),
  template = system.file("templates/compare-runs.qmd", package = "guildai"),
  viewer = getOption("guildai.viewer"),
  ...
)
}
\arguments{
\item{id1, id2}{run ids}

\item{output_dir}{where to place the rendered html}

\item{template}{report template}

\item{viewer}{Viewer to display training run information within (default
to an internal page viewer if available, otherwise to the R session
default web browser, `utils::browseURL()`).}

\item{...}{passed on to `quarto::quarto_render()`}
}
\value{
path to the generated html, invisibly
}
\description{
compare runs
}
