% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-generated-cli-opt-wrappers.R
\name{guild_select_cli}
\alias{guild_select_cli}
\title{guild_select_cli
Select a run and shows its ID.}
\usage{
guild_select_cli(
  ...,
  all = FALSE,
  min = NULL,
  max = NULL,
  short_id = FALSE,
  attr = NULL,
  path = FALSE,
  filter = NULL,
  operation = NULL,
  label = NULL,
  unlabeled = FALSE,
  tag = NULL,
  comment = NULL,
  marked = FALSE,
  unmarked = FALSE,
  started = NULL,
  digest = NULL,
  running = FALSE,
  completed = FALSE,
  error = FALSE,
  terminated = FALSE,
  pending = FALSE,
  staged = FALSE
)
}
\arguments{
\item{...}{passed on to the \code{guild} executable. Arguments are automatically quoted with \code{shQuote()}, unless they are protected with \code{I()}. Pass \code{'--help'} or \code{help = TRUE} to see all options.}

\item{all}{(bool) Select all matching runs, not just the latest.}

\item{min}{Select the run with the lowest value for the specified COLSPEC.}

\item{max}{Select the run with the highest value for the specified COLSPEC.}

\item{short_id}{(bool) Use short ID.}

\item{attr}{Show specified run attribute rather than run ID.}

\item{path}{(bool) Show run path.}

\item{filter}{Filter runs using a filter expression. See Filter by Expression above for details..}

\item{operation}{Filter runs with operations matching \code{VAL}.}

\item{label}{Filter runs with labels matching VAL. To show unlabeled runs, use \code{unlabeled}.}

\item{unlabeled}{(bool) Filter only runs without labels.}

\item{tag}{Filter runs with TAG.}

\item{comment}{Filter runs with comments matching VAL.}

\item{marked}{(bool) Filter only marked runs.}

\item{unmarked}{(bool) Filter only unmarked runs.}

\item{started}{Filter only runs started within RANGE. See above for valid time ranges.}

\item{digest}{Filter only runs with a matching source code digest.}

\item{running}{(bool) Filter only runs that are still running.}

\item{completed}{(bool) Filter only completed runs.}

\item{error}{(bool) Filter only runs that exited with an error.}

\item{terminated}{(bool) Filter only runs terminated by the user.}

\item{pending}{(bool) Filter only pending runs.}

\item{staged}{(bool) Filter only staged runs.}
}
\description{
This command is generally used when specifying a run ID for
another Guild command. For example, to restart the latest \code{train}
run:
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{`guild run --restart $(guild select -o train)`
}\if{html}{\out{</div>}}
\subsection{Specify a Run}{

You may specify a run using a run ID, a run ID prefix, or a
one-based index corresponding to a run returned by the \code{list}
command.

If RUN isn't specified, the latest matching run is selected.
}

\subsection{Selecting Min or Max Scalar}{

To select the run with the lowest or highest column value, use
\code{min} or \code{max} respectively. For example, to select the run
with the lowest \code{loss} scalar value, use \verb{min loss}.

For help with COLSPEC formatting, see \verb{COLUMN SPECS} in \code{compare}
help by running \verb{guild compare help}.

Other run filters are applied before selecting a minimum or
maximium scalar value.
}

\subsection{Filter by Operation}{

Runs may be filtered by operation using \code{operation}.  A run is
only included if any part of its full operation name, including
the package and model name, matches the value.

Use \code{operation} multiple times to include more runs.
}

\subsection{Filter by Label}{

Use \code{label} to only include runs with labels containing a
specified value. To select runs that do not contain a label,
specify a dash '-' for \code{VAL}.

Use \code{label} multiple times to include more runs.
}

\subsection{Filter by Tag}{

Use \code{tag} to only include runs with a specified tag. Tags must
match completely and are case sensitive.

Use \code{tag} multiple times to include more runs.
}

\subsection{Filter by Marked and Unmarked}{

Use \code{marked} to only include marked runs.

Use \code{unmarked} to only include unmarked runs. This option may
not be used with \code{marked}.
}

\subsection{Filter by Source Code Digest}{

To show runs for a specific source code digest, use \code{-g} or
\code{digest} with a complete or partial digest value.
}
}
