% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guild-calls.R
\name{guild_view}
\alias{guild_view}
\title{Launch Guild Viewer}
\usage{
guild_view(..., wait = FALSE)
}
\arguments{
\item{...}{passed on to the `guild` binary}

\item{wait}{}
}
\description{
Launch Guild Viewer
}
