% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guild-bin.R
\name{install_guild}
\alias{install_guild}
\title{Install guildai core}
\usage{
install_guild(guildai = "guildai", python = find_python())
}
\arguments{
\item{guildai}{character vector passed directly to `pip install`.}

\item{python}{path to python binary, used to create a private venv.}
}
\value{
path to the guild binary
}
\description{
Install guildai core
}
\examples{
# install_guild(c("-e", "~/guild/guildai"))
}
