% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_guild_cli}
\alias{install_guild_cli}
\title{Install guild for usage in the Terminal}
\usage{
install_guild_cli(dest = "~/bin", completions = TRUE)
}
\arguments{
\item{dest}{Where to place the \code{guild} executable.
This should be a location on the \code{PATH}.}

\item{completions}{Whether to also install shell completion helpers.}
}
\value{
path to the installed guild executable, invisibly.
}
\description{
This function makes available the \code{guild} executable installed by \code{install_guild()}
for usage in the Terminal.
}
