% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{ls_runs}
\alias{ls_runs}
\title{list guild runs}
\usage{
ls_runs(...)
}
\arguments{
\item{...}{additional arguments passed to \verb{guild api runs}. Try
\code{"--help"} to see options.}
}
\value{
a dataframe of runs
}
\description{
list guild runs
}
\examples{
\dontrun{
withr::with_package("dplyr", {

## sort and filter using scalars
pluck_scalar <- function(scalar_summary_df, val, tag, prefix = ".guild") {
  if(!length(scalar_summary_df)) return(NA)
  out <- scalar_summary_df \%>\%
    filter(tag == {{ tag }}, prefix == {{ prefix }}) \%>\%
    pull({{ val }})
  if(length(out) < 1)
    out <- NA
  out
}

runs <- ls_runs()
runs \%>\%
  rowwise() \%>\%
  mutate(
    min_val_loss =
      pluck_scalar(scalars, min_val, "epoch_loss", "logs/validation"),
    .after = label) \%>\%
  ungroup() \%>\%
  slice_min(min_val_loss, n = 5)

})
}
}
