% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{ls_scalars}
\alias{ls_scalars}
\title{Get full set of runs scalars}
\usage{
ls_scalars(runs = NULL, ...)
}
\arguments{
\item{runs}{a runs selection}

\item{...}{passed on go `guild()`}
}
\value{
a tibble with runs scalars
}
\description{
Get full set of runs scalars
}
\examples{
\dontrun{
ls_scalars(1) # scalars from most recent run
ls_scalars(1:2) # scalars form two most recent runs

# pass in a dataframe of runs
ls_runs() \%>\%
  filter(flags$epochs > 5) \%>\%
  ls_scalars()
}
}
