% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-run.R
\name{run_report}
\alias{run_report}
\title{Render a Run Summary Report}
\usage{
run_report(
  id,
  output_dir = file.path(tempdir(), id),
  template = system.file("templates/run-summary.qmd", package = "guildai"),
  ...
)
}
\arguments{
\item{id}{run id}

\item{output_dir}{directory where to place the rendered document}

\item{template}{path to a paramaterized quarto document}
}
\description{
Render a Run Summary Report
}
