% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{runs_copy}
\alias{runs_copy}
\alias{runs_move}
\title{Move or copy runs}
\usage{
runs_copy(runs = NULL, location, ..., copy_resources = FALSE)

runs_move()
}
\arguments{
\item{runs}{A runs selection}

\item{location}{A directory where to place the runs}

\item{...}{passed on to guild}

\item{copy_resources}{whether run resources should be also copied. If
FALSE, (the default), run resources in the run directory will be
symlinks to a guild managed storage location.}
}
\value{
NULL, invisibly
}
\description{
Move or copy runs
}
