% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{runs_delete}
\alias{runs_delete}
\alias{runs_purge}
\title{Delete runs}
\usage{
runs_delete(runs = NULL, ...)

runs_purge(runs = NULL, ...)
}
\arguments{
\item{runs}{a runs selection}

\item{...}{passed on to `guild()`}
}
\description{
`runs_delete()` moves runs into a guild managed "trash" directory.
`runs_purge()` permanently delete runs from "trash" directory.
  Only deleted runs can be purged.
}
