% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{runs_export}
\alias{runs_export}
\alias{runs_import}
\title{Move or copy runs}
\usage{
runs_export(runs = NULL, location, ..., move = FALSE, copy_resources = FALSE)

runs_import(runs = NULL, location, ..., move = FALSE, copy_resources = FALSE)
}
\arguments{
\item{runs}{A runs selection}

\item{location}{A directory where to place the runs, or find the runs.}

\item{...}{passed on to guild}

\item{move}{bool, whether the runs should be moved or copied by the import or export operation.}

\item{copy_resources}{whether run resources should be also copied. If
\code{FALSE}, (the default), run resources in the run directory will be
symlinks to a guild managed storage location.}
}
\value{
The value supplied to the \code{runs} argument, invisibly.
}
\description{
Move or copy runs
}
