% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{runs_info}
\alias{runs_info}
\title{list guild runs}
\usage{
runs_info(..., full = FALSE, scalars = FALSE)
}
\arguments{
\item{...}{additional arguments passed to `guild api runs`. Try
`"--help"` to see options.}

\item{full}{Whether to include all the available run info.}

\item{scalars}{Wheter to export all scalars for the requested runs instead.}
}
\value{
a dataframe of runs
}
\description{
list guild runs
}
