% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{runs_scalars}
\alias{runs_scalars}
\title{Get Full set of runs scalars}
\usage{
runs_scalars(...)
}
\arguments{
\item{...}{passed on go `guild()`}
}
\value{
a tibble with runs scalars
}
\description{
Get Full set of runs scalars
}
