% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_run_report}
\alias{view_run_report}
\title{Render a Run Summary Report}
\usage{
view_run_report(
  id = ls_runs("1")$id,
  output_dir = file.path(tempdir(), id),
  template = system.file("templates/view-run.qmd", package = "guildai"),
  viewer = getOption("guildai.viewer"),
  ...
)
}
\arguments{
\item{id}{run id. Defaults to the latest run.}

\item{output_dir}{directory where to place the rendered document.}

\item{template}{path to a parameterized quarto document.}

\item{viewer}{Viewer to display training run information within (default
to an internal page viewer if available, otherwise to the R session
default web browser, \code{utils::browseURL()}).}

\item{...}{passed on to \code{quarto::quarto_render()}}
}
\value{
path to the generated html, invisibly
}
\description{
Render a Run Summary Report
}
