% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{write_run_attr}
\alias{write_run_attr}
\title{Write run attributes}
\usage{
write_run_attr(name, data)
}
\arguments{
\item{name}{A string}

\item{data}{The data to write. This needs to be encodable as yaml.
Passing NULL delete the existing attr.}
}
\value{
the written yaml, invisibly.
}
\description{
This function does nothing if \code{is_run_active()} is \code{FALSE}.
}
