% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{export_guild_cli}
\alias{export_guild_cli}
\title{Export guild for usage in the Terminal}
\usage{
export_guild_cli(dest = "~/bin", completions = TRUE)
}
\arguments{
\item{dest}{Where to place the \code{guild} executable.
This should be a location on the PATH.}

\item{completions}{Whether to also install shell completion helpers.}
}
\value{
path to the installed guild executable, invisibly
}
\description{
Export guild for usage in the Terminal
}
