% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-generated-cli-doc-stubs.R
\name{guild_api_runs_cli}
\alias{guild_api_runs_cli}
\title{guild_api_runs_cli
Show runs as JSON.}
\usage{
guild_api_runs_cli(
  ...,
  format = NA,
  filter = NULL,
  operation = NULL,
  label = NULL,
  unlabeled = NA,
  tag = NULL,
  comment = NULL,
  marked = NA,
  unmarked = NA,
  started = NULL,
  digest = NULL,
  running = NA,
  completed = NA,
  error = NA,
  terminated = NA,
  pending = NA,
  staged = NA,
  api_version = NULL,
  deleted = NA,
  include_batch = NA
)
}
\arguments{
\item{...}{passed on to the \code{guild} executable. Arguments are automatically quoted with \code{shQuote()}, unless they are protected with \code{I()}. Pass \code{'--help'} or \code{help = TRUE} to see all options.}

\item{format}{(bool) Format the JSON outout.}

\item{filter}{Filter runs using a filter expression. See Filter by Expression above for details..}

\item{operation}{Filter runs with operations matching \code{VAL}.}

\item{label}{Filter runs with labels matching VAL. To show unlabeled runs, use \code{unlabeled}.}

\item{unlabeled}{(bool) Filter only runs without labels.}

\item{tag}{Filter runs with TAG.}

\item{comment}{Filter runs with comments matching VAL.}

\item{marked}{(bool) Filter only marked runs.}

\item{unmarked}{(bool) Filter only unmarked runs.}

\item{started}{Filter only runs started within RANGE. See above for valid time ranges.}

\item{digest}{Filter only runs with a matching source code digest.}

\item{running}{(bool) Filter only runs that are still running.}

\item{completed}{(bool) Filter only completed runs.}

\item{error}{(bool) Filter only runs that exited with an error.}

\item{terminated}{(bool) Filter only runs terminated by the user.}

\item{pending}{(bool) Filter only pending runs.}

\item{staged}{(bool) Filter only staged runs.}

\item{api_version}{API version.}

\item{deleted}{(bool) Show deleted runs.}

\item{include_batch}{(bool) Include batch runs.}
}
\description{
IMPORTANT: This command is experimental and subject to change without
notice.
}
