% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guild-calls.R
\name{guild_run}
\alias{guild_run}
\title{Launch a guild run}
\usage{
guild_run(opspec = "train.R", flags = NULL, wait = TRUE, echo = wait, ...)
}
\arguments{
\item{opspec}{typically path to an R script, but could be an scalar
string that guild recognizes as a valid opspec.}

\item{flags}{- a named list or vector like `c(noise = .3, dropout = .4)`

  - a scalar string like `"noise=.3 dropout=.4"`

  - a dataframe of flags for a batch of runs}

\item{wait}{whether to wait for the run to finish}

\item{echo}{whether output from the run is shown in the current R
console. Note, this has no effect on whether expressions are echoed in
the guild run stdout.}

\item{...}{passed through to [base::system2]}
}
\value{
NULL, invisibly. This function is called for its side effect.
}
\description{
Launch a guild run
}
