% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guild.R
\name{guild_run}
\alias{guild_run}
\title{Launch a guild run}
\usage{
guild_run(
  opspec = "train.R",
  flags = NULL,
  ...,
  background = FALSE,
  label = NULL,
  tags = NULL
)
}
\arguments{
\item{opspec}{typically path to an R script, but could be an scalar
string that guild recognizes as a valid opspec.}

\item{flags}{\itemize{
\item a named list or vector like \code{c(noise = .3, dropout = .4)}
\itemize{
\item a scalar string like \code{"noise=.3 dropout=.4"}
\item a dataframe of flags for a batch of runs
}
}}

\item{...}{
  Arguments passed on to \code{\link[base:system2]{base::system2}}
  \describe{
    \item{\code{command}}{the system command to be invoked, as a character string.}
    \item{\code{args}}{a character vector of arguments to \command{command}.}
    \item{\code{stdout,stderr}}{where output to \file{stdout} or
    \file{stderr} should be sent.  Possible values are \code{""}, to the \R
    console (the default), \code{NULL} or \code{FALSE} (discard output),
    \code{TRUE} (capture the output in a character vector) or a
    character string naming a file.}
    \item{\code{stdin}}{should input be diverted?  \code{""} means the default,
    alternatively a character string naming a file.  Ignored
    if \code{input} is supplied.}
    \item{\code{input}}{if a character vector is supplied, this is copied one
    string per line to a temporary file, and the standard input of
    \code{command} is redirected to the file.}
    \item{\code{env}}{character vector of name=value strings to set environment
    variables.}
    \item{\code{wait}}{a logical (not \code{NA}) indicating whether the \R
    interpreter should wait for the command to finish, or run it
    asynchronously.  This will be ignored (and the interpreter will
    always wait) if \code{stdout = TRUE} or \code{stderr = TRUE}. When
    running the command asynchronously, no output will be displayed on
    the \code{Rgui} console in Windows (it will be dropped, instead).}
    \item{\code{timeout}}{timeout in seconds, ignored if 0.  This is a limit for the
    elapsed time running \code{command} in a separate process.   Fractions
    of seconds are ignored.}
    \item{\code{minimized,invisible}}{arguments that are accepted on Windows but
    ignored on this platform, with a warning.}
  }}

\item{background}{whether to do the run in the background. If \code{TRUE},
\code{guild_run()} returns immediately.}

\item{label, tags}{optional strings used to label or tag experiments.}
}
\value{
the return value from \code{system2()}, invisibly. This function is
primarily called for its side effect.
}
\description{
Launch a guild run
}
