% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_guild}
\alias{install_guild}
\title{Install guildai core}
\usage{
install_guild(
  guildai = "https://api.github.com/repos/guildai/guildai/tarball/HEAD",
  python = find_python()
)
}
\arguments{
\item{guildai}{character vector passed directly to \verb{pip install}. To
install the release version of guildai, this can be \code{"guildai"}.}

\item{python}{path to python binary, used to create a private venv.}
}
\value{
path to the guild binary
}
\description{
This installs the \code{guild} executable for use by the package.
}
\examples{
# install_guild(c("-e", "~/guild/guildai"))
# install_guild("~/guildai", reticulate::install_python())
# install_guild("https://api.github.com/repos/guildai/guildai/tarball/HEAD")
# install_guild(
#   guildai = "https://api.github.com/repos/guildai/guildai/tarball/HEAD",
#   python = reticulate::install_python())
#
}
