% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_guild}
\alias{install_guild}
\title{Install guildai core}
\usage{
install_guild(
  guildai = "https://api.github.com/repos/guildai/guildai/tarball/HEAD",
  python = find_python()
)
}
\arguments{
\item{guildai}{Character vector of arguments passed directly to \verb{pip install}. To install the release version of guildai, this can be
\code{"guildai"}.}

\item{python}{Path to a python binary, used to create a private venv.}
}
\value{
path to the \code{guild} executable
}
\description{
This installs the \code{guild} executable for use by the R package. It creates
an isolated python virtual environment private to the R package and installs
guildai into it. Repeated calls to \code{install_guild()} result in a
fresh installation.
}
\details{
It requires that a suitable python version is
available on the system.
}
\examples{
## Install release version:
# install_guild("guildai", python = reticulate::install_python())

## Install development version
# install_guild(
#   guildai = "https://api.github.com/repos/guildai/guildai/tarball/HEAD",
#   python = reticulate::install_python("3.10:latest"))

## Install local development version:
# install_guild(c("-e", "~/guild/guildai"))
}
