% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guild-calls.R
\name{ls_runs}
\alias{ls_runs}
\title{list guild runs}
\usage{
ls_runs(..., full = FALSE)
}
\arguments{
\item{...}{additional arguments passed to `guild api runs`. Try
`"--help"` to see options.}

\item{full}{Whether to include all the available run info.}
}
\value{
a dataframe of runs
}
\description{
list guild runs
}
