% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{runs_label}
\alias{runs_label}
\alias{runs_tag}
\alias{runs_mark}
\title{Annotate runs}
\usage{
runs_label(runs, ..., action = c("append", "prepend", "set", "remove"))

runs_tag()

runs_mark()
}
\arguments{
\item{runs}{a runs selection}

\item{...}{passed on to `guild()`}

\item{action}{what action to take w/ existing labels/tags}

\item{label, tag}{string}
}
\description{
Annotate runs
}
