% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{runs_label}
\alias{runs_label}
\alias{runs_tag}
\alias{runs_mark}
\title{Annotate runs}
\usage{
runs_label(runs = NULL, label, ...)

runs_tag(
  runs = NULL,
  tag,
  ...,
  action = c("add", "remove", "clear"),
  sync_labels = FALSE
)

runs_mark()
}
\arguments{
\item{runs}{a runs selection}

\item{label, tag}{string}

\item{...}{passed on to \code{guild()}}

\item{action}{what action to take respective to existing tags}
}
\description{
Annotate runs
}
