% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{runs_label}
\alias{runs_label}
\alias{runs_tag}
\alias{runs_mark}
\alias{runs_comment}
\title{Annotate runs}
\usage{
runs_label(runs = NULL, label, ..., clear = is.null(label))

runs_tag(runs = NULL, add = NULL, ..., remove = NULL, clear = FALSE)

runs_mark(runs = NULL, ..., clear = FALSE)

runs_comment(runs = NULL, comment = NULL, ..., clear = FALSE)
}
\arguments{
\item{runs}{a runs selection}

\item{label, comment}{a string}

\item{...}{passed on to \code{guild}}

\item{clear}{bool, whether to clear the existing tags/comments first.}

\item{add, remove}{a character vector of tags to add or remove}
}
\description{
Annotate runs
}
\examples{
if(FALSE) {

ls_runs(1) \%>\% runs_tag(clear = TRUE)
ls_runs(1) \%>\% runs_tag("foo")
ls_runs(1)$tags
ls_runs(1) \%>\% runs_tag("bar")
ls_runs(1)$tags
ls_runs(1) \%>\% runs_tag(remove = "foo")
ls_runs(1)$tags
ls_runs(1) \%>\% runs_tag("baz", clear = TRUE)
ls_runs(1)$tags

## pass through options to `guild tag` cli subcommand
runs_tag(NULL, NULL, "--help")
ls_runs(1) \%>\% runs_tag("--add" = c("foo", "bar"))
ls_runs(1) \%>\% runs_tag("--add" = "baz", "--delete" = "bar")
ls_runs(1)$tags

}
}
